autojs/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "autojs"
version = "0.1.0"
description = "jSeeker Automation — auto-apply engine, ATS runners, answer bank"
requires-python = ">=3.11"
dependencies = [
    "jseeker",
    "playwright",
    "pydantic>=2.0",
    "pyyaml",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["autojs*"]
